cleaning up, have to find a way to assign opp overload with const members

+++ mod02/ex02/Fixed.cpp
@@ -6,9 +6,12 @@
 /*   By: zenotan <zenotan@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/15 01:37:55 by zenotan       #+#    #+#                 */
-/*   Updated: 2021/10/18 22:13:51 by zenotan       ########   odam.nl         */
+/*   Updated: 2021/10/18 23:26:58 by zenotan       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "Fixed.hpp"
+#include <cmath>
 
 #include "Fixed.hpp"
 #include <cmath>
@@ -136,14 +139,14 @@ Fixed		Fixed::operator-(const Fixed &num) const
 Fixed		Fixed::operator/(const Fixed &num) const
 {
 	Fixed	cpy;
-	cpy._raw_bits = _raw_bits / num.toInt();
+	cpy._raw_bits = _raw_bits / num.toFloat();
 	return cpy;
 }
 
 Fixed		Fixed::operator*(const Fixed &num) const
 {
 	Fixed	cpy;
-	cpy._raw_bits = _raw_bits * num.toInt();
+	cpy._raw_bits = _raw_bits * num.toFloat();
 	return cpy;
 }
 
@@ -199,4 +202,4 @@ const Fixed&  Fixed::max(const Fixed &num_a, const Fixed &num_b)
 	if (num_a._raw_bits > num_b._raw_bits)
 		return num_a;
 	return num_b;
-}	
+}
